test: add tests for XMLParser security configuration
